Update tests using APIs removed in newer Symfony versions
