+++ tests/Rules/Symfony/console_application_loader.php
@@ -6,5 +6,5 @@ use Symfony\Component\Console\Application;
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
 $application = new Application();
-$application->add(new ExampleCommand());
+$application->addCommands([new ExampleCommand()]);
 return $application;
