+++ tests/Type/Symfony/data/ExampleAbstractControllerWithoutContainer.php
@@ -10,21 +10,6 @@ use function PHPStan\Testing\assertType;
 final class ExampleAbstractControllerWithoutContainer extends AbstractController
 {
 
-	public function services(): void
-	{
-		assertType('object', $this->get('foo'));
-		assertType('object', $this->get('synthetic'));
-		assertType('object', $this->get('bar'));
-		assertType('object', $this->get(doFoo()));
-		assertType('object', $this->get());
-
-		assertType('bool', $this->has('foo'));
-		assertType('bool', $this->has('synthetic'));
-		assertType('bool', $this->has('bar'));
-		assertType('bool', $this->has(doFoo()));
-		assertType('bool', $this->has());
-	}
-
 	public function parameters(ContainerInterface $container, ParameterBagInterface $parameterBag): void
 	{
 		assertType('array|bool|float|int|string|null', $container->getParameter('unknown'));
